conference: restructure message processing
- Simplify function names a bit (we don't necessary need to mention DC
  in them, it seems like it does not really give much of a context as
  the functino implementation is agnostic from the fact where the
  message was received from: over the data channel, Matrix signaling or
  any other thing).
- Remove separate file that contained processing of data channel only
  messages as all of them are strictly speaking regular peer messages,
  it seems like currently grouping these 100+ LOC together in a separate
  file does not simplify things since it only forces us to jump forwards
  and backwards without separating things (they are still methods on a
  same structure).
